A ton of changes were made due to the code coverage reports that helped identified areas where the unit tests were not correctly implemented or insufficient.
Code coverage is now:
Instruction - 70%
Line - 70%
Branch - 50%
================
AVERAGE: 79.17%
================

+++ DiscoveryAccountSystemDomain/src/main/java/za/ac/nwu/domain/dto/ExchangeMediumDto.java
@@ -119,7 +119,7 @@ public class ExchangeMediumDto implements Serializable {
         this.Date = date;
     }
 
-    public ExchangeMediumDto(int exchangeMediumID, String type, String description, int balance, LocalDate date, int id) {
+    public ExchangeMediumDto(Integer exchangeMediumID, String type, String description, double balance, LocalDate date, Integer id) {
         this.exchangeMediumID = exchangeMediumID;
         this.Type = type;
         this.Description = description;
